вынес типы в university.types.ts + добавил подстраницу с фидбеком для препода
